m0leconteaserctf2025/wolfram/site/requirements.txt
argon2-cffi
flask
flask-login
flask-sqlalchemy
flask-wtf
gunicorn
